+++ src/mainwindow.cpp
@@ -9110,10 +9110,47 @@ void MainWindow::moveSelectedTagsToTagId(int tagId) {
     }
 
     if (tagList.count() > 0) {
+        const bool useScriptingEngine =
+            ScriptingService::instance()->noteTaggingHookExists();
+        QVector<Note> notes;
+
+        // workaround when signal block doesn't work correctly
+        directoryWatcherWorkaround(true, true);
+
+        if (useScriptingEngine) {
+            notes = Note::fetchAll();
+        }
+
         // move tags
         for (Tag tag : Utils::asConst(tagList)) {
+            if (useScriptingEngine) {
+                const QList<Tag> tagsToHandle =
+                    Tag::fetchRecursivelyByParentId(tag.getId());
+
+                // check all tags we need to handle
+                for (const Tag &tagToHandle : tagsToHandle) {
+                    // remove tag from all notes
+                    for (const Note &note : notes) {
+                        handleScriptingNoteTagging(note, tagToHandle, true, false);
+                    }
+                }
+            }
+
             tag.setParentId(tagId);
             tag.store();
+
+            if (useScriptingEngine) {
+                const QList<Tag> tagsToHandle =
+                    Tag::fetchRecursivelyByParentId(tag.getId());
+
+                // check all tags we need to handle
+                for (const Tag &tagToHandle : tagsToHandle) {
+                    // add tag to all notes
+                    for (const Note &note : notes) {
+                        handleScriptingNoteTagging(note, tagToHandle, false, false);
+                    }
+                }
+            }
 
             showStatusBarMessage(
                 tr("Moved tag '%1' to new tag").arg(tag.getName()), 3000);
@@ -9121,6 +9158,9 @@ void MainWindow::moveSelectedTagsToTagId(int tagId) {
 
         reloadCurrentNoteTags();
         reloadTagTree();
+
+        // turn off the workaround again
+        directoryWatcherWorkaround(false, true);
     }
 }